# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "KIN API",
    "version": "1.0.0",
    "description": "KIN AI Companion Platform — We Build You A Friend\n\nThe KIN API powers Mission Control (web dashboard), Telegram/Discord/WhatsApp bots, local KIN instances, and third-party integrations. It provides:\n\n- **Auth** — Telegram Login Widget + JWT-based sessions\n- **Chat** — Multi-companion AI chat with a two-brain supervisor (Local Ollama → Groq → Anthropic/OpenAI fallback)\n- **Skills Marketplace** — Browse, install, and custom-submit companion skill extensions\n- **Companion Skills** — NFT-portable XP/leveling system for companion skill trees\n- **Heartbeat & Recovery** — Local KIN health monitoring and snapshot-based recovery\n- **Billing** — Stripe subscription checkout, portal, and webhook handling\n- **Support** — AI-powered support chat with human escalation, FAQ, tickets, and feature requests\n\nAll protected routes require a JWT bearer token obtained via `POST /auth/telegram` or `POST /auth/dev-login` (development only).",
    "contact": {
      "name": "KIN Support",
      "url": "https://www.meetyourkin.com/support",
      "email": "support@meetyourkin.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/kr8tiv-ai/kr8tiv-runtime-truth-contracts/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    },
    {
      "url": "https://api.meetyourkin.com",
...
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token. Obtain via `POST /auth/telegram` or `POST /auth/dev-login`. Token expires after 2 hours. Algorithm: HS256."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "oneOf": [
              { "type": "string" },
              {
                "type": "object",
                "properties": {
                  "message": { "type": "string" },
                  "statusCode": { "type": "integer" }
                }
              }
            ]
          }
        },
        "example": { "error": "Unauthorized" }
      },
...
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "user-abc123" },
          "telegramId": { "type": "integer", "example": 123456789 },
          "username": { "type": "string", "nullable": true, "example": "alice" },
          "firstName": { "type": "string", "example": "Alice" },
          "lastName": { "type": "string", "nullable": true, "example": "Smith" },
          "tier": {
            "type": "string",
            "enum": ["free", "pro", "enterprise"],
            "example": "pro"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-01-15T09:00:00.000Z"
          },
          "onboardingComplete": { "type": "boolean", "example": true }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": ["token", "user"],
        "properties": {
          "token": {
            "type": "string",
            "description": "JWT bearer token. Valid for 2 hours.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": { "$ref": "#/components/schemas/UserProfile" }
        }
      },
...
    },
    "responses": {
...
      "TooManyRequests": {
        "description": "Rate limit exceeded. Default: 100 requests/minute per IP in production.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "statusCode": { "type": "integer", "example": 429 },
                "error": { "type": "string", "example": "Too Many Requests" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [],
...
  "paths": {
...
    "/auth/telegram": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authTelegram",
        "summary": "Telegram Login Widget authentication",
        "description": "Verifies the Telegram Login Widget payload using HMAC-SHA256 with the bot token secret key. Creates a new user account on first login. Returns a signed JWT valid for 2 hours.\n\nThe `hash` field must be a valid Telegram login hash and the `auth_date` must be within 24 hours.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id", "first_name", "auth_date", "hash"],
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "Telegram user ID.",
                    "example": 123456789
                  },
                  "first_name": {
                    "type": "string",
                    "description": "User's first name from Telegram.",
                    "example": "Alice"
                  },
                  "last_name": {
                    "type": "string",
                    "description": "User's last name (optional).",
                    "example": "Smith"
                  },
                  "username": {
                    "type": "string",
                    "description": "Telegram username without @ (optional).",
                    "example": "alice_smith"
                  },
                  "photo_url": {
                    "type": "string",
                    "format": "uri",
                    "description": "Profile photo URL (optional)."
                  },
                  "auth_date": {
                    "type": "integer",
                    "description": "Unix timestamp of authentication. Must be within 24 hours.",
                    "example": 1704067200
                  },
                  "hash": {
                    "type": "string",
                    "description": "HMAC-SHA256 verification hash computed by Telegram.",
                    "example": "a3f9d2c1..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "401": {
            "description": "Invalid Telegram hash or expired auth_date.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Invalid Telegram authentication" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" },
          "500": {
            "description": "Server misconfiguration — bot token not set.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/auth/verify": {
      "get": {
        "tags": ["Auth"],
        "operationId": "authVerify",
        "summary": "Verify JWT token",
        "description": "Decodes the bearer token and returns the associated user profile. Returns `{valid: false}` for expired or malformed tokens without raising an error — safe to poll from the client.",
        "responses": {
          "200": {
            "description": "Token verification result.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "valid": { "type": "boolean", "enum": [true] },
                        "userId": { "type": "string" },
                        "user": { "$ref": "#/components/schemas/UserProfile" }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "valid": { "type": "boolean", "enum": [false] }
                      }
                    }
                  ]
                }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
    "/auth/dev-login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authDevLogin",
        "summary": "Dev-only login (development environment only)",
        "description": "Creates or finds a user by Telegram ID without hash verification. **Only available when `NODE_ENV=development`.** This endpoint does not exist in production. Useful for local testing and smoke tests.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["telegramId", "firstName"],
                "properties": {
                  "telegramId": {
                    "type": "integer",
                    "description": "Any Telegram-style numeric ID to use as test identity.",
                    "example": 999001
                  },
                  "firstName": {
                    "type": "string",
                    "description": "Display name for the test user.",
                    "example": "DevUser"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          },
          "404": {
            "description": "Endpoint not found (production environment).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
  }
}